// types/schema.d.ts
export interface components {
    schemas: {
// ...
        LoanResponse: {
            /** Loan Amount */
            loan_amount: string;
            /** Interest Rate */
            interest_rate: string;
            /** Duration Months */
            duration_months: number;
            /** Id */
            id: number;
            /** User Id */
            user_id: number;
            /** Monthly Repayment */
            monthly_repayment: string;
            /** Total Repayable */
            total_repayable: string;
            /** Amount Paid */
            amount_paid: string;
            /** Balance */
            balance: string;
            status: components["schemas"]["LoanStatus"];
            /**
             * Application Date
             * Format: date-time
             */
            application_date: string;
            /** Approval Date */
            approval_date: string | null;
            /** Disbursement Date */
            disbursement_date: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
        SavingsPaymentResponse: {
            /** Id */
            id: number;
            /** User Id */
            user_id: number;
            /** Amount */
            amount: string;
            type: components["schemas"]["SavingsPaymentTypeSchema"];
            /**
             * Payment Date
             * Format: date-time
             */
            payment_date: string;
            /** Payment Month */
            payment_month: string | null;
            /** Description */
            description: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
        };
// ...
        ShareResponse: {
            /** Shares Count */
            shares_count: number;
            /** Share Value */
            share_value: string;
            /** Id */
            id: number;
            /** User Id */
            user_id: number;
            /** Total Value */
            total_value: string;
            /**
             * Purchase Date
             * Format: date-time
             */
            purchase_date: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}